requirements.txt
black
flake8
httpx
pre-commit
python-dotenv
rope
